Created CategorySlider
